kymo: move pixel calibration out of kymo file
